requirements.txt
streamlit>=1.35.0
pandas>=2.2
plotly>=5.22
gspread>=6.1
google-auth>=2.29
streamlit-extras==0.4.0 
python-dotenv
streamlit-autorefresh